+++ Igloo/Assets/Code/Scripts/Dungeon/Algorithm/NpcAction/QuizNpc/Quiz3Npc.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Quiz3Npc : AlgoNpcAction
+{
+
+
+    protected override void ActionStart()
+    {
+
+    }
+}
