+++ Igloo/Assets/Code/Scripts/Dungeon/Algorithm/NpcAction/QuizNpc/Quiz3Npc.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 22482864aa5752740bb006fe92e4dc49
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
