+++ Igloo/Assets/Code/Scripts/Dungeon/Algorithm/NpcAction/QuizNpc/QuizNpcGivingKey.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 10e49f80aba6fd34cb6aeb754b471853
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
