+++ Igloo/Assets/Prefabs/Dungeon/Algorithm/DungeonItem/FinalKey.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b8bd913624abd88408fc52fc8a46a366
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
