+++ Igloo/Assets/Resources/AlgoDialogue.csv
@@ -1,26 +1,76 @@
-ID,이름,대사,선택지 유무,선택지 대사,선택 후 대사,대화 종료 여부,종료 후 이벤트
-1,맹송이,안녕' 여기서 뭐해?,1,,,0,0
-,,,,길을 잃었어.,2,,0
-,,,,그냥 구경 중이야.,2,,0
-,,,,너 혹시 알고리즘 던전으로 가는 길을 아니?,4,,0
-2,맹송이,여기서 이어지는 길은 알고리즘 던전으로 가는 길뿐이야.,0,,,0,0
-3,맹송이,그 길이라도 알려줄까?,1,,,0,0
-,,,,응! 알려줘!,5,,0
-,,,,아니' 거긴 가고 싶지 않아.,9,,0
-4,맹송이,아는데 알려줄까?,1,,,0,0
-,,,,응! 알려줘!,5,,0
-,,,,아직 갈지말지 고민중이야...,9,,0
-5,맹송이,맨입으로? 내가 지금 고민중인 문제가 있는데... ,0,,,0,0
-,,답을 맞히면 생각해볼게.,1,,,0,0
-,,,,무슨 문젠데?,6,,0
-,,,,와; 그냥 혼자 찾을게.,9,,0
-6,맹송이,간단해.,0,,,0,0
-,,1~100의 숫자 중에 내가 생각하는 숫자를 가장 빨리 맞히는 방법이 뭘까?,0,,,0,0
-,,네가 고른 숫자보다 큰지 작은지는 말해줄 수 있어.,1,,,0,0
-,,,,나는 언제나 운이 좋아. 아무거나 때려맞히면 돼!!@!,8,,0
-,,,,1부터 순차적으로 확인해보면 어떨까?,8,,0
-,,,,반씩 후보를 줄여나가면 어떨까? 50부터 시작하는거지!,7,,0
-7,맹송이,너 꽤나 똑똑하구나? ,0,,,0,0
-,,이걸 가지고 가면 될거야.,0,,,1,1
-8,맹송이,우.....,0,,,1,0
-9,맹송이,아.. 그렇구나..,0,,,1,0
+ID,이름,대사,선택지 유무,선택지 대사,선택 후 대사,대화 종료 여부,종료 후 이벤트,선택지 활성화 여부
+1,맹송이,안녕! 여기서 뭐해?,1,,,0,0,0
+,,,,길을 잃었어.,2,0,0,0
+,,,,그냥 구경 중이야.,2,0,0,0
+,,,,너 혹시 알고리즘 던전으로 가는 길을 아니?,4,0,0,0
+2,맹송이,여기서 이어지는 길은 알고리즘 던전으로 가는 길뿐이야.,0,,,0,0,0
+3,맹송이,그 길이라도 알려줄까?,1,,,0,0,0
+,,,,응! 알려줘!,5,0,0,0
+,,,,아니' 거긴 가고 싶지 않아.,9,0,0,0
+4,맹송이,아는데 알려줄까?,1,,,0,0,0
+,,,,응! 알려줘!,5,0,0,0
+,,,,아직 갈지말지 고민중이야...,9,0,0,0
+5,맹송이,맨입으로? 내가 지금 고민중인 문제가 있는데... ,0,,,0,0,0
+,,답을 맞히면 생각해볼게.,1,,,0,0,0
+,,,,무슨 문젠데?,6,0,0,0
+,,,,와; 그냥 혼자 찾을게.,9,0,0,0
+6,맹송이,간단해.,0,,,0,0,0
+,,1~100의 숫자 중에 내가 생각하는 숫자를 가장 빨리 맞히는 방법이 뭘까?,0,,,0,0,0
+,,네가 고른 숫자보다 큰지 작은지는 말해줄 수 있어.,1,,,0,0,0
+,,,,나는 언제나 운이 좋아. 아무거나 때려맞히면 돼!!@!,8,0,0,0
+,,,,1부터 순차적으로 확인해보면 어떨까?,8,0,0,0
+,,,,반씩 후보를 줄여나가면 어떨까? 50부터 시작하는거지!,7,0,0,0
+7,맹송이,너 꽤나 똑똑하구나? ,0,,,0,0,0
+,,이걸 가지고 가면 될거야.,0,,,1,1,0
+8,맹송이,우.....,0,,,1,0,0
+9,맹송이,아.. 그렇구나..,0,,,1,0,0
+10,문지기,거길 들어갔다 무사히 나온 송이는 본 적이 없어... ,0,,,0,0,0
+,,정말 들어갈거야?,1,,,0,0,0
+,,,,응! 들어가볼게!,11,0,0,0
+,,,,더 효율적인 방법이 있다면 .. 들어가고 싶지 않아,12,0,0,0
+11,문지기,행운을 빌어. 이걸 가지고 가면 조금이나마 도움이 될거야.,0,,,1,1,0
+12,문지기 ,나를 좀 도와주면 나도 너를 도와줄 수 있을 것 같은데 들어볼래?,1,,,0,0,0
+,,,,아니야! 난 꼭 들어가보고 싶어! 무사히 돌아올게!,11,0,0,0
+,,,,응 뭔데?,13,0,0,0
+13,문지기,친구들 사이에서 유행인 카드 게임이 있는데 필승법을 찾고 있어!!,0,,,0,0,0
+,,우선 각각 다른 레벨을 가진 카드들이 순서대로 나열되어있어.,0,,,0,0,0
+,,인접한 카드를 합성해 업그레이드할 수 있고,0,,,0,0,0
+,,한 번 업그레이드된 레벨은 더 이상 올라가지 않아.,0,,,0,0,0
+,,합성 할때 두 카드 레벨의 합만큼의 골드를 얻을 수 있어. ,0,,,0,0,0
+,,이때 최대치의 골드를 얻을 수 있는 방법이 뭐가 있을까?,1,,,0,0,0
+,,,,반씩 나눠 생각해보자! ,14,0,0,1
+,,,,작은 단위로 쪼개 생각해보자. ,14,0,0,1
+,,,,최대 레벨을 가진 카드를 먼저 선택해 합성해야해!,15,0,0,1
+14,문지기 ,아닌 것 같아... 상부상조는 못 할 것 같다. 각자도생하자.,0,,,1,0,0
+15,문지기 ,오!! 그렇게 해볼게. 고마워! 보답으로 ...,0,,,0,0,0
+,,돌아갈 수 있는 길을 알려줄게.,0,,,0,0,0
+,,문제에는 언제나 다양한 길이 있는 법이지!,0,,,1,1,0
+16,콩송이,나 좀 도와줄래?,1,,,0,0,0
+,,,,여기는 도움이 필요한 친구들이 참 많다. 네 문제는 뭔데?,17,0,0,0
+,,,,미안.. 바빠서,18,0,0,0
+17,콩송이,고마워! 내가 요새 알바를 하는데 문제가 생겨서 말이야..,0,,,0,0,0
+,,점장님이 지폐 말고 동전으로만 거슬러주라고 하지 뭐야!,0,,,0,0,0
+,,500원 100원 50원 10원의 동전들을 가지고,0,,,0,0,0
+,,가장 적은 동전 개수를 사용해 거스름돈을 주는 방법이 뭐가 있을까? ,1,,,0,0,0
+,,,,잡히는 대로 아무거나 주면 되지 뭐 이런 걸 고민해.,19,0,0,0
+,,,,작게 나누어 생각해보자!,19,0,0,0
+,,,,큰 동전들부터 주는 방식으로 생각해봐!,20,0,0,0
+18,콩송이,그래. 안녕. ,0,,,1,0,0
+19,콩송이,너무 오래 걸릴 것 같아! 그렇게 일하다간 혼날 것 같은데...?,0,,,1,0,0
+20,콩송이,오 유용한 조언같아! ,0,,,0,0,0
+,,당장에 가장 빨리 거스름돈을 채울 수 있는 선택을 하면 된다는 거구나!,0,,,0,0,0
+,,고마워. 답례로 이걸 줄게.,0,,,1,1,0
+21,팥송이,0 1 앞의 두 개를 더하면 1. 그 다음은 2.. 그 다음은 3.. 그 다음은 5..,1,,,0,0,0
+,,,,저기 뭐해?,22,0,0,0
+22,팥송이,아.. 너 혹시 피보나치 수열 알아?,0,,,0,0,0
+,,피보나치 수열의 n번째를 계산하는 프로그램을 짜고 있는데.. ,0,,,0,0,0
+,,재귀함수를 호출하니 시간이 너무 오래 걸리는 거 있지? ,0,,,0,0,0
+,,어떻게 하면 좋을지 생각중이었어..,1,,,0,0,0
+,,,,그렇구나. 열심히 해.,23,0,0,0
+,,,,재귀함수가 끝날때까지 기다리자. 그냥!!!,24,,,0
+,,,,작게 나눠 생각하고 그 값을 저장해두고 쓰면 어때?,25,,,0
+23,팥송이,그래. 안녕. ,0,,,1,0,0
+24,팥송이,그 방법은 아니야. ,0,,,1,0,0
+25,팥송이,오! n 번째 수를 구하기 위해 n-1번째와 n-2가 쓰이니,0,,,0,0,0
+,,이것들을 기억해두고 재사용하라는거구나! ,0,,,0,0,0
+,,고마워!,0,,,1,1,0
