--- Igloo/Assets/Resources/AlgoDialogueExtendTest.csv.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5a76628cdbcf06b41a1874e5f538565d
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
