+++ Igloo/Assets/Resources/AlgoDialogueFinal.csv.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: dcdb9999bf38a1f4bb91ec9d6cedd02d
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
