--- Igloo/Assets/Resources/AlgoDialogueV3.csv
@@ -1,42 +1,0 @@
-ID,이름,대사,선택지 유무,선택지 대사,선택 후 대사,대화 종료 여부,종료 후 이벤트
-1,맹송이,안녕' 여기서 뭐해?,1,,,0,0
-,,,,길을 잃었어.,2,0,0
-,,,,그냥 구경 중이야.,2,0,0
-,,,,너 혹시 알고리즘 던전으로 가는 길을 아니?,4,0,0
-2,맹송이,여기서 이어지는 길은 알고리즘 던전으로 가는 길뿐이야.,0,,,0,0
-3,맹송이,그 길이라도 알려줄까?,1,,,0,0
-,,,,응! 알려줘!,5,0,0
-,,,,아니' 거긴 가고 싶지 않아.,9,0,0
-4,맹송이,아는데 알려줄까?,1,,,0,0
-,,,,응! 알려줘!,5,0,0
-,,,,아직 갈지말지 고민중이야...,9,0,0
-5,맹송이,맨입으로? 내가 지금 고민중인 문제가 있는데... ,0,,,0,0
-,,답을 맞히면 생각해볼게.,1,,,0,0
-,,,,무슨 문젠데?,6,0,0
-,,,,와; 그냥 혼자 찾을게.,9,0,0
-6,맹송이,간단해.,0,,,0,0
-,,1~100의 숫자 중에 내가 생각하는 숫자를 가장 빨리 맞히는 방법이 뭘까?,0,,,0,0
-,,네가 고른 숫자보다 큰지 작은지는 말해줄 수 있어.,1,,,0,0
-,,,,나는 언제나 운이 좋아. 아무거나 때려맞히면 돼!!@!,8,0,0
-,,,,1부터 순차적으로 확인해보면 어떨까?,8,0,0
-,,,,반씩 후보를 줄여나가면 어떨까? 50부터 시작하는거지!,7,0,0
-7,맹송이,너 꽤나 똑똑하구나? ,0,,,0,0
-,,이걸 가지고 가면 될거야.,0,,,1,1
-8,맹송이,우.....,0,,,1,0
-9,맹송이,아.. 그렇구나..,0,,,1,0
-10,문지기,거길 들어갔다 무사히 나온 송이는 본 적이 없어... ,0,,,0,0
-,,정말 들어갈거야?,1,,,0,0
-,,,,응! 들어가볼게!,11,0,0
-,,,,더 효율적인 방법이 있다면 .. 들어가고 싶지 않아,12,0,0
-11,문지기,행운을 빌어. 이걸 가지고 가면 조금이나마 도움이 될거야.,0,,,1,1
-12,문지기 ,나를 좀 도와주면 나도 너를 도와줄 수 있을 것 같은데 들어볼래?,1,,,0,0
-,,,,아니야! 난 꼭 들어가보고 싶어! 무사히 돌아올게!,11,0,0
-,,,,응 뭔데?,13,0,0
-13,문지기,문제 문제 문제 문제 문제 문제 문제 문제 문제,1,,,0,0
-,,,,오답 오답 오답,14,0,0
-,,,,오답 오답 오답,14,0,0
-,,,,정답 정답 정답,15,0,0
-14,문지기 ,아닌 것 같아... 상부상조는 못 할 것 같다. 각자도생하자.,0,,,1,0
-15,문지기 ,오!! 그렇게 해볼게. 고마워! 보답으로 ...,0,,,0,0
-,,돌아갈 수 있는 길을 알려줄게.,0,,,0,0
-,,문제에는 언제나 다양한 길이 있는 법이지!,0,,,1,1
