--- Igloo/Assets/Resources/AlgoDialogueV3.csv.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 6c95f3dbff3e892488d412c5bf9b504e
-TextScriptImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
