+++ Igloo/Assets/Resources/Items/DungeonItem/AlgoDungeon/FinalKey.asset
@@ -1,0 +1,19 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dce8d3218704dbb42bb00e132f89c4e4, type: 3}
+  m_Name: FinalKey
+  m_EditorClassIdentifier: 
+  id: 205
+  itemName: "\uD070 \uC5F4\uC1E0"
+  value: 0
+  icon: {fileID: 21300000, guid: b6ce924f699b771488503cc5801c5dfd, type: 3}
+  itemDesc: "\uC5B4\uB518\uAC00\uC758 \uD070 \uBB38\uC744 \uC5F4 \uC218 \uC788\uB2E4."
