+++ Igloo/Assets/Resources/Items/DungeonItem/AlgoDungeon/FinalKey.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7203b01c021f19d43a44f1790e23c4ba
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
