+++ Igloo/Assets/Resources/Items/DungeonItem/AlgoDungeon/TempItem1.asset
@@ -13,7 +13,7 @@ MonoBehaviour:
   m_Name: TempItem1
   m_EditorClassIdentifier: 
   id: 203
-  itemName: tempitem
+  itemName: "\uB3C4\uB11B"
   value: 0
-  icon: {fileID: 21300000, guid: e4f36d8907c172f4bb249a911cb69364, type: 3}
+  icon: {fileID: 21300000, guid: 702fb173ee6c9144198697ad8b1138ac, type: 3}
   itemDesc: "\uC9C0\uCE5C \uB208\uC1A1\uC774\uAC00 \uC88B\uC544\uD560 \uAC83 \uAC19\uB2E4!"
