+++ Igloo/Assets/Scenes/Quests/Algorithm_2.unity
@@ -66466,7 +66466,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 9078484029274698496, guid: c26000b352498d5409066aa54fd50499, type: 3}
       propertyPath: csv_FileName
-      value: AlgoDialogueV3
+      value: AlgoDialogue
       objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
