+++ Igloo/Assets/Timeline/CutsceneAG.playable.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 62ef0e3a1312a51448d98aff3f6b4028
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
