+++ Igloo/UserSettings/EditorUserSettings.asset
@@ -6,34 +6,34 @@ EditorUserSettings:
   serializedVersion: 4
   m_ConfigSettings:
     RecentlyUsedSceneGuid-0:
-      value: 5455505507060b0d5a0f542045755b44134f4e7e7c2e276675704c36e3e6673a
-      flags: 0
-    RecentlyUsedSceneGuid-1:
-      value: 0753515f52015a5f545e5a7711710b4443164a782a2970672b284435e6e2366d
-      flags: 0
-    RecentlyUsedSceneGuid-2:
-      value: 5508055e5407580e550d5f7411710b44144f1d2b7d7f773379784835bab7666f
-      flags: 0
-    RecentlyUsedSceneGuid-3:
-      value: 575352075d545f5d585e5e2147765c444e161b2e7d2b73642b284464bbb0673d
-      flags: 0
-    RecentlyUsedSceneGuid-4:
-      value: 0208060403560d02580b0d7b40775a15471641727d2d73677f791960e6b6666f
-      flags: 0
-    RecentlyUsedSceneGuid-5:
-      value: 0753065e5200515d5d0a5d2711770f44124f4d797d7a70647e781f64b3b36c6c
-      flags: 0
-    RecentlyUsedSceneGuid-6:
-      value: 065703030101580c580f5a2641715c4445164a7e7d2c7e687d7a4861e6e1613e
-      flags: 0
-    RecentlyUsedSceneGuid-7:
       value: 555403500702590b580f5d27442506441516402f7e782534757e1e32e7b1356f
       flags: 0
-    RecentlyUsedSceneGuid-8:
+    RecentlyUsedSceneGuid-1:
+      value: 5755055107010f0e5c5c0d7647770f411516497e7d2a22327e284a37e7b6633d
+      flags: 0
+    RecentlyUsedSceneGuid-2:
       value: 07080c565d560f5a545b0d7414200844411649782e2927622c7b1f36bae46469
       flags: 0
+    RecentlyUsedSceneGuid-3:
+      value: 065703030101580c580f5a2641715c4445164a7e7d2c7e687d7a4861e6e1613e
+      flags: 0
+    RecentlyUsedSceneGuid-4:
+      value: 5653525001070b5f085b082448740e44144e1b787c2e743328711c37e7b83061
+      flags: 0
+    RecentlyUsedSceneGuid-5:
+      value: 5602015257075e5d090a0a2714720b444216417f292a7035782a4430b7e1633c
+      flags: 0
+    RecentlyUsedSceneGuid-6:
+      value: 0208060403560d02580b0d7b40775a15471641727d2d73677f791960e6b6666f
+      flags: 0
+    RecentlyUsedSceneGuid-7:
+      value: 07065207560c5a085b0d0a2012760c44134e4e28747a71687e7f4932b5e4316f
+      flags: 0
+    RecentlyUsedSceneGuid-8:
+      value: 02545152060758095c5a587b47740e4415154c7d7b707463797a1b30b6b8653c
+      flags: 0
     RecentlyUsedSceneGuid-9:
-      value: 5755055107010f0e5c5c0d7647770f411516497e7d2a22327e284a37e7b6633d
+      value: 575352075d545f5d585e5e2147765c444e161b2e7d2b73642b284464bbb0673d
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
